refactor: updated import to prevent circle import error
